RestoreSystemProperties sets correct properties.
RestoreSystemProperties now replaces the system properties with a copy
of the original properties. This ensures that the properties behave like
the original properties. That was not the case before. E.g. the
`contains` method returned `false` for a property that has been
originally set but was not modified in the test.

Fixes #43.