Variables can be set after creating the rule
Some environment variables have the same value for all tests. Someone
who reads the test can easily recognize this when the set and clear
methods for these variables are directly called within the statement
that creates the EnvironmentVariables rule.

    @Rule
    public final EnvironmentVariables environmentVariables
        = new EnvironmentVariables()
            .set("name", "value")
            .clear("another variable");

This also fixes a bug where the environment variables are not reset when
the set and clear methods are called from within the static initializer
of the EnvironmentVariables rule. E.g.

    @Rule
    public final EnvironmentVariables environmentVariables
        = new EnvironmentVariables() {{
            set("dummy variable", "dummy value"
        }};

Fixes #49 and #61.