Use random Strings as values if possible
Improve quality of the test. Unfortunately the tests are not completely
independent because they use the global environment variables map.
Therefore variables that we set in a test may leak to another test. By
using a random value we reduce the likelihood that a test is successful
although the expected value of a variable was set by another test.