Fix building with OpenJDK 6 in Travis CI
OpenJDK 6 is no longer supported in the default build environment of
Travis CI, so downgrade to Ubuntu 14.04 LTS (Trusty Tahr). This version
is EOL (like Java 6), but seems to be the only way to continue building
with OpenJDK 6. Even with the downgraded build environment, a manual
workaround is necessary since the usual "jdk" key does not work.